python-lab-16/python-lab-16.1/requirements.txt
python-dotenv>=1.0.1
gigachat>=0.1.35
